requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
flake8==6.1.0
idna==3.6
mccabe==0.7.0
pycodestyle==2.11.1
pyflakes==3.1.0
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
requests==2.31.0
tinydb==4.8.0
urllib3==2.1.0
